+++ Models/Entity.cs
@@ -11,13 +11,13 @@ public partial class Entity
 
     public string? Caption { get; set; }
 
-    public short EntityTypeId { get; set; }
+    public int EntityTypeId { get; set; }
 
     public short Status { get; set; }
 
     public string Symbol { get; set; } = null!;
 
-    public short MaterialId { get; set; }
+    public int MaterialId { get; set; }
 
-    public short EntityMode { get; set; }
+    public int EntityMode { get; set; }
 }
